brian_multiprocess_udp.py
Using the class BrianConnectUDP it is possible to run SNN simulations using multiple processes or even multiple machines and all the spikes are going to be exchanged using UDP packages. The idea was to create something to help with the interaction between SNN simulators and the real world.
This system always tries to run in real time and if necessary it will drop packets. So this is not a step-by-step simulation where all spikes sent by a node are to be received by another one. The order the packets arrive is non deterministic too as well the time.